requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
bitarray==2.5.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
crc16==0.1.1
crc32c==2.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
django-filter==22.1
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
drf-yasg==1.21.3
duplicity==0.8.21
ed25519==1.5
fasteners==0.14.1
filelock==3.7.1
frozenlist==1.3.0
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.7.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.0.1
mnemonic==0.20
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
PAM==0.4.2
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
phonenumbers==8.12.52
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
PyGObject==3.42.0
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-decouple==3.6
python3-openid==3.2.0
pyTON==0.1.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
QtPy==2.0.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
systemd-python==234
tvm-valuetypes==0.0.9
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.4.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uvloop==0.16.0
virtualenv==20.15.1
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
